+++ src/index.ts
@@ -100,7 +100,7 @@ export { LanguageCode, LanguageMap, NullableLanguageMap } from "./client/CachedA
 export { EnkaClientOptions } from "./client/EnkaClient";
 export { SubstatsContainer } from "./models/artifact/Artifact";
 export { EquipType } from "./models/artifact/ArtifactData";
-export { DynamicData } from "./models/assets/DynamicTextAssets";
+export { DynamicData, GenshinPlatform, UserInfo } from "./models/assets/DynamicTextAssets";
 export { BodyType, Gender, CharacterRarity } from "./models/character/CharacterData";
 export { Birthday, CharacterVoiceActors, VoiceLanguage } from "./models/character/CharacterDetails";
 export { GameServerRegion } from "./models/enka/EnkaUser";
@@ -109,6 +109,7 @@ export { WeaponType } from "./models/weapon/WeaponData";
 export { ElementType } from "./models/Element";
 export { FightProp } from "./models/StatProperty";
 export { CharacterPreview } from "./models/User";
+export { Paths } from "./utils/ts_utils";
 
 // functions
 export { fetchJSON } from "./utils/axios_utils";
@@ -116,4 +117,5 @@ export { validateCache } from "./utils/cache_utils";
 export { getCharactersById, getNameIdByCharacterId, hasEnergySkill, isReleased } from "./utils/character_utils";
 
 // constants
+export { defaultEnkaClientOptions } from "./client/EnkaClient";
 export { artifactRarityRangeMap, fightProps, percent } from "./utils/constants";