+++ src/models/character/talents/Skill.ts
@@ -3,6 +3,7 @@ import EnkaClient from "../../../client/EnkaClient";
 import AssetsNotFoundError from "../../../errors/AssetsNotFoundError";
 import ImageAssets from "../../assets/ImageAssets";
 import TextAssets from "../../assets/TextAssets";
+import DynamicTextAssets from "../../assets/DynamicTextAssets";
 
 /**
  * Normal Attack, Elemental Skill, and Elemental Burst. Not including Passive Talents.
@@ -16,7 +17,7 @@ class Skill {
     /**  */
     readonly name: TextAssets;
     /**  */
-    readonly description: TextAssets;
+    readonly description: DynamicTextAssets;
     /**  */
     readonly icon: ImageAssets;
 
@@ -36,7 +37,7 @@ class Skill {
 
         this.name = new TextAssets(json.getAsNumber("nameTextMapHash"), enka);
 
-        this.description = new TextAssets(json.getAsNumber("descTextMapHash"), enka);
+        this.description = new DynamicTextAssets(json.getAsNumber("descTextMapHash"), {}, enka);
 
         this.icon = new ImageAssets(json.getAsString("skillIcon"), enka);
     }